fix: rename tests
