+++ examples/user-account/src/user_account_repository.rs
@@ -1,5 +1,5 @@
 use crate::user_account::{UserAccount, UserAccountEvent, UserAccountId};
-use event_store_adapter_rs::event_store::EventStoreForDynamoDB;
+use event_store_adapter_rs::event_store_for_dynamodb::EventStoreForDynamoDB;
 use event_store_adapter_rs::types::{Aggregate, EventStore, EventStoreReadError, EventStoreWriteError};
 
 pub struct UserAccountRepository {
