+++ lib/src/event_store_for_dynamodb.rs
