+++ lib/src/lib.rs
@@ -1,6 +1,6 @@
-pub mod event_store;
+pub mod event_store_for_dynamodb;
 #[cfg(test)]
-mod event_store_test;
+mod event_store_for_dynamodb_test;
 pub mod key_resolver;
 pub mod serializer;
 pub mod types;
